+++ src/main/java/nextstep/app/FormLoginInterceptor.java
@@ -1,27 +1,31 @@
-package nextstep.app.ui;
+package nextstep.app;
 
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
+import nextstep.app.ui.AuthenticationException;
+import org.springframework.web.servlet.HandlerInterceptor;
 
 import java.util.Map;
 
-@RestController
-public class LoginController {
+public class FormLoginInterceptor implements HandlerInterceptor {
+
     public static final String SPRING_SECURITY_CONTEXT_KEY = "SPRING_SECURITY_CONTEXT";
 
     private final MemberRepository memberRepository;
 
-    public LoginController(MemberRepository memberRepository) {
+    public FormLoginInterceptor(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
     }
 
-    @PostMapping("/login")
-    public ResponseEntity<Void> login(HttpServletRequest request, HttpSession session) {
+    @Override
+    public boolean preHandle(
+            HttpServletRequest request,
+            HttpServletResponse response,
+            Object handler
+    ) {
         Map<String, String[]> parameterMap = request.getParameterMap();
         String username = parameterMap.get("username")[0];
         String password = parameterMap.get("password")[0];
@@ -30,8 +34,9 @@ public class LoginController {
                 .filter(it -> it.matchPassword(password))
                 .orElseThrow(AuthenticationException::new);
 
+        HttpSession session = request.getSession();
         session.setAttribute(SPRING_SECURITY_CONTEXT_KEY, member);
 
-        return ResponseEntity.ok().build();
+        return false;
     }
 }
